requirements.txt
pytesseract~=0.3.13
pyTelegramBotAPI~=4.27.0
torch~=2.7.0
pillow~=11.0.0
pix2tex~=0.1.4